% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClimNOAA.R
\name{getClimNOAA}
\alias{getClimNOAA}
\title{getClimNOAA: Download monthly gridded NOAA climate data}
\usage{
getClimNOAA(
  park = "all",
  year = as.integer(format(Sys.Date(), "\%Y")),
  months = 5
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD monitored for water quality.}
\item{"ACAD"}{Acadia NP only}
\item{"BOHA"}{Boston Harbor only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{year}{4-digit year to query. Earliest available year is 1895, and latest is 2024. Currently can only
handle 1 year at a time.}

\item{months}{Vector of numeric months to query. Typically there's about a 6 week delay in monthly data
availability.}
}
\value{
Data frame of NOAA monthly gridded climate data for each specified site.
}
\description{
This function downloads monthly gridded climate data directly from NOAA's National Centers
for Environmental Information (www.ncei.noaa.gov) for the centroid of each park in NETN. The returned
data frame includes total precipitation (prcp in mm/day), Maximum air temperature (tmax in C) averaged
over the month, minimum air temperature (tmin in C) averaged over the month, and average temperature
(tavg in C). Note that occasionally you're unable to connect to the server, and will receive an error message
when that happens. This function was used to compile historic weather data for each park from 1895 up
through May 2024, which is stored as a package dataset called NETN_clim_annual (accessed via
\code{data("NETN_clim_annual")}). You should only need this function to download newer months of data
as they become available.
}
\examples{
\dontrun{

# get climate data for January - May 2024 for all NETN (will take a bit to download)
getClimNOAA(months = 1:5, year = 2024)

# get climate data for May 2025 for ROVA
getClimNOAA(park = "ROVA", months = 5, year = 2025)

# get climate data for all of 2023 in LNETN
getClimNOAA(park = "LNETN", year = 2023, months = 1:12)

}

}
