% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClimBar.R
\name{plotClimBar}
\alias{plotClimBar}
\title{plotClimBar: Plot climate trends as bar plot}
\usage{
plotClimBar(
  park = "all",
  years = 2006:format(Sys.Date(), "\%Y"),
  months = 5:10,
  parameter = NA,
  units = "sci",
  facet_park = FALSE,
  facet_param = FALSE,
  palette = "#1e90ff",
  numcol = 2,
  x_pad = c(0.01, 0),
  legend_position = "none",
  gridlines = "none"
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD monitored for water quality.}
\item{"ACAD"}{Acadia NP only}
\item{"BOHA"}{Boston Harbor only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{years}{Numeric. Years to query. Accepted values start at 1895.}

\item{months}{Numeric. Months to query by number. Accepted values range from 1:12.
Default is 5:10. If specifying new months not yet included in NETN_clim_annual dataset, will
download months that are available from NOAA.}

\item{parameter}{Specify the parameter(s) to plot. Acceptable values are
\describe{
\item{"all"}{Plot all climate variables}
\item{"ppt"}{Monthly total precipitation.}
\item{"tmax"}{Monthly average maximum temperature.}
\item{"tmin}{Monthly average minimum temperature.}
\item{"tmean}{Monthly average temperature.}
}}

\item{units}{Specify if you want Scientific or English units. Acceptable values are "sci" (default) and "eng".
If "sci", temperature units are in C and precipitation units are in mm. If "eng", temperature units are in F,
and precipitation units are in inches.}

\item{facet_park}{Logical. If TRUE, plots sites on separate facets (ie figures). If FALSE (Default),
plots all sites on the same figure. This is only enabled if multiple sites are chosen.}

\item{facet_param}{Logical. If TRUE, plots parameters on separate facets. If FALSE (Default), plots
all parameters on the same figure. Note that function will automatically facet if parameters with
different units (ie tmean and ppt) are selected.}

\item{palette}{Specify a vector of colors. Default is light blue ('#1e90ff') If fewer colors than
parameters are specified, they will be ramped to generate enough colors.}

\item{numcol}{Specify number of columns in the facet wrap, which is only enabled when either multiple years
are specified or multiple parks. Default is 2.}

\item{x_pad}{Vector of 2 values. Specifies how much padding to add to x-axis on either side of axis.
Default is c(0.01, 0)}

\item{legend_position}{Specify location of legend. To turn legend off, use legend_position =
"none" (Default). Other options are "top", "bottom", "left", "right".}

\item{gridlines}{Specify whether to add gridlines or not. Options are c("none" (Default),
"grid_y", "grid_x", "both")}

\item{line_width}{Numeric. Specify the width of the line. Default is 1.}
}
\value{
Returns a ggplot object of specified monthly climate trends as bar chart
}
\description{
This function produces bar plot filtered on park, year, month, and climate parameter.
This function was developed to compare with monthly precip. with monthly discharge measurements for
water monitoring, and by default only plots May - October. If you're plotting all months, use plotClimTrend()
instead of this function. If multiple parks are specified, they can either be plotted on the same figure or
separate figures. If multiple parameters are specified, they will be plotted on separate figures.
}
\examples{
\dontrun{

# Plot total monthly precip for 2006:2023 for May - Oct in English units
plotClimBar(park = "MABI", years = 2006:2023, parameter = "ppt", units = 'eng')


}

}
