% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClimCumPrecip.R
\name{plotClimCumPrecip}
\alias{plotClimCumPrecip}
\title{plotClimCumPrecip: Plot cumulative precip relative to historic}
\usage{
plotClimCumPrecip(
  park = "all",
  years = format(Sys.Date(), "\%Y"),
  layers = "lines",
  normal = "norm20cent",
  units = "sci",
  palette = "#1e90ff",
  plot_title = TRUE,
  title_type = "UnitCode",
  legend_position = "right",
  numcol = 3,
  gridlines = "none",
  include_error = TRUE,
  legend_row = 1
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD monitored for water quality.}
\item{"ACAD"}{Acadia NP only}
\item{"BOHA"}{Boston Harbor only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{years}{Numeric. Years to plot separately. Accepted values start at 2006.If multiple years
specified, will facet results on year. Default is current year.}

\item{layers}{Options are "points", "lines", "bars", or "area" for Annual Values.
By default, only lines and points will plot. Bars or area should be selected on their own.}

\item{normal}{Specify normals to plot. By default, the 20th century normal (1901-2000) plots.
Other options include:
\describe{
\item{"norm20cent"}{Plots the 20th century normal (1901 - 2000)}
\item{"norm1990"}{Plots the 30-year norm from 1991 - 2020}
}}

\item{units}{Specify if you want Scientific or English units. Acceptable values are "sci" (default) and "eng".
If "sci" precipitation units are mm; if "eng", precipitation units are in inches.}

\item{palette}{Color palette for years. If only 1 color is specified and multiple years
are specified, all will be the same color.}

\item{plot_title}{Logical. If TRUE (default) prints park name at top of figure. If FALSE,
does not print site name. Only enabled when one park is selected.}

\item{title_type}{Specify whether to label facets with 4-letter UnitCode (default) or full UnitName.
Options are c("UnitCode", "UnitName").}

\item{legend_position}{Specify location of legend. To turn legend off, use legend_position =
"none" (Default). Other options are "top", "bottom", "left", "right".}

\item{numcol}{Specify number of columns in the facet wrap, which is only enabled when either multiple years
are specified or multiple parks. Default is 3.}

\item{gridlines}{Specify whether to add gridlines or not. Options are c("none" (Default), "grid_y", "grid_x", "both")}

\item{include_error}{Logical. If TRUE (default), plots min/max, 95\% and 50\% error bands around the normal.
If FALSE, only plots the normal line.}

\item{legend_row}{Integer. Specify number of rows to plot the legend. Default is 1.}
}
\value{
Returns a ggplot object of cumulative precipitation trends
}
\description{
This function plots cumulative monthly precipitation for a given year compared
to either the 19th century or 30-year normal. If multiple parks or years are specified, resulting
plots will be faceted on those variables.
}
\examples{
\dontrun{

# Plot ACAD cumulative precipitation for 2020 through 2023 with gridlines on y axis and precip in inches
plotClimCumPrecip(park = "ACAD", years = 2020:2023, legend_position = 'bottom',
                 gridlines = "grid_y", units = 'eng')

# Plot ACAD cumulative precipitation for 2014 through 2024 with legend 2 rows and points and lines
plotClimCumPrecip(park = "ACAD", years = 2014:2024, legend_position = 'bottom', legend_row = 2,
layers = c("points", "lines"))

# Plot all but SAIR cumulative precipitation for 2023 with 4 columns
parks <- c("ACAD", "MABI", "MIMA", "MORR", "ROVA", "SAGA", "SARA", "WEFA")
plotClimCumPrecip(park = parks, years = 2023, legend_position = 'bottom', numcol = 4)


}

}
