% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClimDrought.R
\name{plotClimDrought}
\alias{plotClimDrought}
\title{plotClimDrought: Plots weekly drought based on county-level drought index}
\usage{
plotClimDrought(
  park = "all",
  years = as.numeric(format(Sys.Date(), format = "\%Y")),
  months = 1:12,
  dom_county = TRUE,
  end_date = NA,
  legend_position = "bottom",
  plot_title = TRUE,
  gridlines = "none",
  x_pad = c(0, 0),
  legend_row = 1
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD monitored for water quality.}
\item{"ACAD"}{Acadia NP only}
\item{"BOHA"}{Boston Harbor only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{years}{Vector of years to download drought index for, will start with the first Tuesday of the year
and end with the last Tuesday of the year specified. Defaults to current year.}

\item{months}{Vector of months to download drought index for. Default = 1:12.}

\item{dom_county}{Logical. If TRUE (Default) only plots predominant county if park covers multiple counties.
If FALSE, facets on county.}

\item{end_date}{Quoted end date of query. If blank (Default), will return the most recent week of data available.}

\item{legend_position}{Specify location of legend. To turn legend off, use legend_position = "none". Other
options are "top", "bottom (default)", "left", "right".}

\item{plot_title}{Logical. If TRUE (default) prints site name at top of figure. If FALSE,
does not print site name. Only enabled when one site is selected.}

\item{gridlines}{Specify whether to add gridlines or not. Options are c("none" (Default), "grid_y", "grid_x", "both")}

\item{x_pad}{Vector of 2 values. Specifies how much padding to add to x-axis on either side of axis.
Default is c(0, 0)}

\item{legend_row}{Integer. Specify number of rows to plot the legend. Default is 1.}
}
\value{
Returns a ggplot object of specified drought trends
}
\description{
This function plots weekly drought index at the county level for each specified park.
Resulting plot shows the percent of county area in 5 levels of drought, with D0 = Abnormally Dry,
D1 = Moderate Drought, D2 = Severe Drought, D3 = Extreme Drought, and D4 = Exceptional Drought.
Drought designations and color palette come from the U.S. Drought Monitor. If multiple parks are
specified, results will be faceted with a separate plot for each park. Where multiple counties
occur in a park, data can be faceted by county if dom_county = FALSE. To only plot predominant
county, specify dom_county = TRUE (default).
}
\examples{
\dontrun{

# Plot drought info for MABI and SAGA for 2020 to 2024
plotClimDrought(park = c("MABI", "SAGA"), years = 2020:2024)

# Plot drought info for both MORR counties for 2023
plotClimDrought(park = "MORR", years = 2023, dom_county = FALSE)

# Plot drought info for ACAD dominant county from 2006:2026
plotClimDrought(park = "ACAD", years = 2006:2026)
}

}
