% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClimRel.R
\name{plotClimRel}
\alias{plotClimRel}
\title{plotClimRel: Plot climate data relative to average value}
\usage{
plotClimRel(
  park = "all",
  years = 2006:format(Sys.Date(), "\%Y"),
  months = 1:12,
  normal = "norm20cent",
  parameter = "tmean",
  units = "sci",
  plot_title = TRUE,
  title_type = "UnitCode",
  palette = "viridis",
  color_rev = FALSE,
  legend_position = "right",
  numcol = 3,
  gridlines = "none"
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD monitored for water quality.}
\item{"ACAD"}{Acadia NP only}
\item{"BOHA"}{Boston Harbor only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{years}{Numeric. Years to plot separately. Accepted values start at 2006.If multiple years
specified, will facet results on year.}

\item{normal}{Specify normal to plot. By default, the 20th century normal (1901-2000) plots.
Other options include:
\describe{
\item{"norm20cent"}{Plots the 20th century normal (1901 - 2000)}
\item{"norm1990"}{Plots the 30-year norm from 1991 - 2020}
}}

\item{parameter}{Specify the monthly averaged parameter to plot. Acceptable values are below.
Can only select 1 option.
\describe{
\item{"temp"}{Plot all temperature comparisons.}
\item{"tminmax}{Plot min and max temperature comparisons.}
\item{"tmean"}{Plot mean temperature comparisons.}
\item{"tmax"}{Plot max temperature comparisons.}
\item{"tmin"}{Plot min temperature comparisons.}
\item{"ppt"}{Plot total precipitation comparisons.}
}}

\item{units}{Specify if you want Scientific or English units. Acceptable values are "sci" (default) and "eng".
If "sci", temperature units are in C and precipitation units are in mm. If "eng", temperature units are in F,
and precipitation units are in inches.}

\item{plot_title}{Logical. If TRUE (default) prints site name at top of figure. If FALSE,
does not print site name. Only enabled when one site is selected.}

\item{title_type}{Specify whether to label facets with 4-letter UnitCode (default) or full UnitName.
Options are c("UnitCode", "UnitName").}

\item{palette}{Color palette for plots. Options currently are 'viridis' (yellow - green - blue),
magma (yellow, red, purple), plasma (brighter version of magma), turbo (rainbow),
or create your own by specifying 2 or more colors in quotes, like palette = c("red", "yellow", "blue"),
or palette = c("yellow", "blue"). Hexcodes work too. If only 1 parameter is specified, only need to
specify 1 color.}

\item{color_rev}{Reverse the order of the color pallet. Only enabled for palette = 'viridis'.}

\item{legend_position}{Specify location of legend. To turn legend off, use legend_position =
"none" (Default). Other options are "top", "bottom", "left", "right".}

\item{numcol}{Specify number of columns in the facet wrap, which is only enabled when either multiple years
are specified or multiple parks. Default is 3.}

\item{gridlines}{Specify whether to add gridlines or not. Options are c("none" (Default), "grid_y", "grid_x", "both")}
}
\value{
Returns a ggplot object of specified climate trends
}
\description{
This function plots a given year compared against either the 19th century
normal (1901 - 2000) or the latest 30-year normal (1991 - 2020) using gridded NOAA NClim data
for both the normals and the latest data. Works only at the park level. If a year x month
combination is specified that doesn't occur yet in NETN_clim_annual dataset, it will be
downloaded if available. New months are typically available within a few weeks of the month end.
This function is similar to plotClimAnom, but doesn't color code based on above or below average,
and allows you to include multiple parameters (e.g., tmin and tmax) on the same plot
}
\examples{
\dontrun{

# Plot all temperature variables on 1 graph for MABI in 2024 using English units.
plotClimRel(park = "MABI", years = 2024, parameter = "temp", palette = c("#EEE55A", "#D56062", "#067BC2"),
  units = 'eng')

# Plot precip for SARA 2023 compared to 1990 - 2019 normals.
plotClimRel(park = "SARA", years = 2023, parameter = "ppt", palette = "grey", normal = "norm1990")

# Plot precip for Kroma Kill in SARA 2023 and 2024 compared to 1990 - 2019 normals.
plotClimRel(park = "SARA", years = 2023:2024, parameter = "ppt",
            palette = "grey", normal = "norm1990")

}

}
