% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClimTrend.R
\name{plotClimTrend}
\alias{plotClimTrend}
\title{plotClimTrend: Plot climate trends}
\usage{
plotClimTrend(
  park = "all",
  years = 2006:format(Sys.Date(), "\%Y"),
  months = 1:12,
  layers = c("points", "lines"),
  avg_window = 5,
  parameter = NA,
  units = "sci",
  facet_park = FALSE,
  facet_param = FALSE,
  palette = "viridis",
  span = 0.3,
  plot_se = FALSE,
  numcol = 2,
  x_pad = c(0.01, 0),
  line_width = 1,
  legend_position = "none",
  gridlines = "none"
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD monitored for water quality.}
\item{"ACAD"}{Acadia NP only}
\item{"BOHA"}{Boston Harbor only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{years}{Numeric. Years to query. Accepted values start at 1895.}

\item{months}{Numeric. Months to query by number. Accepted values range from 1:12.
If specifying new months not yet included in NETN_clim_annual dataset, will
download months that are available from NOAA.}

\item{layers}{Options are "points", "lines", "smooth", "rollavg", and "bar". By default, both points
and lines will plot. The lines option connects each monthly value with a linear line. If you include
lines and smooth, the loess-smoothed line will also plot. The rollavg will plot a rolling average (5-year default).
The bar argument plots a bar chart. Note that for precipitation, you can either include points, lines, or bar,
or include rolling average because rolling average is an annual statistic. For temperature parameters,
rolling averages are averaged at the annual level, then the rolling average is calculated. For precipitation,
the annual statistic is total annual precipitation, then summarized with a rolling average.}

\item{avg_window}{Number of years to include in the rolling average, if rollavg specified in layers. Default is 5,
and must be >= 1. Currently only works at annual level, not monthly. Must also use a window that is smaller than
the span of years specified. Preferably this is only used for long time-series.}

\item{parameter}{Specify the parameter(s) to plot. Acceptable values are
\describe{
\item{"all"}{Plot all climate variables}
\item{"ppt"}{Monthly total precipitation.}
\item{"tmax"}{Monthly average maximum temperature.}
\item{"tmin}{Monthly average minimum temperature.}
\item{"tmean}{Monthly average temperature.}
}}

\item{units}{Specify if you want Scientific or English units. Acceptable values are "sci" (default) and "eng".
If "sci", temperature units are in C and precipitation units are in mm. If "eng", temperature units are in F,
and precipitation units are in inches.}

\item{facet_park}{Logical. If TRUE, plots sites on separate facets (ie figures). If FALSE (Default),
plots all sites on the same figure. This is only enabled if multiple sites are chosen.}

\item{facet_param}{Logical. If TRUE, plots parameters on separate facets. If FALSE (Default), plots
all parameters on the same figure. Note that function will automatically facet if parameters with
different units (ie tmean and ppt) are selected.}

\item{palette}{Theme to plot points and lines. Options include 'viridis' (Default- ranges of blue,
green and yellow), magma (yellow, red, purple), plasma (brighter version of magma), turbo (rainbow),
or specify a vector of colors manually. If fewer colors than parameters are specified, they will be
ramped to generate enough colors.}

\item{span}{Numeric. Determines how smoothed the line will be for layers = 'smooth'. Default is 0.3. Higher spans (up to 1)
cause more smoothing (straighter lines). Smaller spans are wavier. Span can range from 0 to 1. Span of 1 is linear.}

\item{plot_se}{Logical. If TRUE, will plot a standard error ribbon. If FALSE (Default), will not plot a ribbon.
Only enabled if layers = "smooth" is specified.}

\item{numcol}{Specify number of columns in the facet wrap, which is only enabled when either multiple years
are specified or multiple parks. Default is 2.}

\item{x_pad}{Vector of 2 values. Specifies how much padding to add to x-axis on either side of axis.
Default is c(0.01, 0)}

\item{line_width}{Numeric. Specify the width of the line. Default is 1.}

\item{legend_position}{Specify location of legend. To turn legend off, use legend_position = "none" (Default). Other
options are "top", "bottom", "left", "right".}

\item{gridlines}{Specify whether to add gridlines or not. Options are c("none" (Default), "grid_y", "grid_x", "both")}
}
\value{
Returns a ggplot object of specified climate trends
}
\description{
This function produces a line or smoothed trend plot filtered on park, year, month, and
climate parameter. If multiple parks are specified, they can either be plotted on the same figure or separate
figures. If multiple parameters are specified, they will be plotted on separate figures.
}
\examples{
\dontrun{

# Plot total monthly precip for 2006:2023 and all months, without smoothing in English units
plotClimTrend(park = "MABI", years = 2006:2023, parameter = "ppt", units = 'eng')

# Plot avg temp in MABI from 1895:2024, without points and with smoothed line and span 0.1.
plotClimTrend(park = "MABI", years = 1895:2024, layers = 'lines', parameter = "tmean", span = 0.1)

# Plot monthly mean temperature for MABI and SARA from 2006:2024, with smoothed line, span 0.7,
and only sample months using the plasma color palette.
plotClimTrend(park = c("MABI", "SARA"), years = 2006:2024,
              parameter = "tmean", span = 0.7, months = 5:10, palette = "plasma")

# Plot 5-year rolling average temp with points for ACAD full time-series
plotClimTrend(park = "ACAD", parameter = "tmean", layers = c("points", "rollavg"), years = 1900:2024)

# Plot 10-year rolling average of all temp params for ACAD full time-series with manual colors
plotClimTrend(park = "ACAD", parameter = c("tmean", "tmax", "tmin"),
  years = 1900:2026,
  layers = "rollavg", avg_window = 10,
  palette = c("black", "red", "blue"))

}

}
