% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_NETN.R
\name{theme_NETN}
\alias{theme_NETN}
\title{theme_NETN: custom ggplot2 theme for climateNETN}
\usage{
theme_NETN()
}
\value{
This function must be used in conjunction with a ggplot object
}
\description{
This function customizes a theme for plotting NETN climate data, including removing the
default panel grids from ggplot2 figures.
}
