% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClimDaymet.R
\name{getClimDaymet}
\alias{getClimDaymet}
\title{getClimDaymet: Download daily Daymet gridded climate data directly from REST}
\usage{
getClimDaymet(park = "all", years = c(2006:2023))
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD monitored for water quality.}
\item{"ACAD"}{Acadia NP only}
\item{"BOHA"}{Boston Harbor only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{years}{Vector of years to download Daymet data. Earliest available year is 1980. Latest is
currently 12/31/2023.}
}
\value{
Data frame of Daymet daily climate data for each specified park.
}
\description{
This function downloads daily gridded climate data directly from Daymet REST for
each NETN park centroid based on its lat/long coordinates, and binds results into a single data frame.
If downloading for all parks and multiple years, function may be slow. The returned data frame
includes Day length (dayl in s/day), Precipitation (prcp in mm/day), Shortwave radiation
(srad in W/m2), Snow water equivalent (swe in kg/m2), Maximum air temperature (tmax in C),
Minimum air temperature (tmin in C), and Water vapor pressure (vp in Pascals). More details on
metrics can be found online: https://daymet.ornl.gov/overview.html > Parameters, Parameter
abbreviations, Units and Descriptions. Note that occasionally you're unable to connect to the
server, and will receive an error message when that happens.
}
\examples{
\dontrun{
# RUN FIRST
library(climateNETN)

# download for MORR 2023 only
morr <- getClimDaymet(park = "MORR", years = 2023)

# download for ACAD from 1980:2023
acad <- getClimDaymet(park = "ACAD", years = 1980:2023)

}

}
