% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClimDrought.R
\name{getClimDrought}
\alias{getClimDrought}
\title{getClimDrought: Download weekly county-level drought index}
\usage{
getClimDrought(
  park = "all",
  years = c(2006:2023),
  week_start = NA,
  dom_county = TRUE
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD monitored for water quality.}
\item{"ACAD"}{Acadia NP only}
\item{"BOHA"}{Boston Harbor only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{years}{Vector of years to download drought index for, will start with 01/01/year and end with 12/31/year.}

\item{week_start}{Quoted start of week formatted as "mm/dd/yyyy". If specified, will return drought index for
given week. If blank, will return drought index for all weeks available in specified years. Note that weekly
data are available every Tuesday. If the week_start specified is not a Tuesday, code will shift the date to the
Tuesday of the specified start week.}

\item{dom_county}{Logical. If TRUE (Default) only plots predominant county if park covers multiple counties.
If FALSE, facets on county.}
}
\value{
Data frame of weekly county drought data for each park.
}
\description{
This function downloads weekly drought index at the county level based on the U.S. Drought Monitor
for each specified park. If downloading for multiple parks and multiple weeks/years, function may be slow.
Returned data frame includes percent of county area in 5 levels of drought, with D0 = Abnormally Dry, D1 =
Moderate Drought, D2 = Severe Drought, D3 = Extreme Drought, and D4 = Exceptional Drought. Also returned is the
Drought Severity and Coverage Index, which ranges from 0 to 500, and is a weighted sum of area within each of
the drought categories. A score of 500 indicates the entire area is in exceptional drought (D4). Where multiple
counties occur in a park, data can be faceted by county if dom_county = FALSE. To only plot predominant county,
specify dom_county = TRUE (default). Otherwise, only park-level data are returned, as the results won't vary
by site within a park. More info on the data are at #' https://droughtmonitor.unl.edu/CurrentMap.aspx.
Note that Drought Monitor reports data such that, when drought a level moves into a higher category (e.g. D1 = 10),
the previous level (e.g. D0) is listed as 100. The columns with pct (e.g. D0pct) correct for that, so that
D0 + ... + D05 + None = 100.
}
\examples{
\dontrun{
# RUN FIRST
library(waterNETN)
importData()

# Get drought info for MABI and SAGA sites for first week of May
mabisaga <- getClimDrought(park = c("MABI", "SAGA"), week_start = "05/01/2024")

# Get drought info for MABI from 2020:2024
mabi5 <- getClimDrought(park = "MABI", years = 2020:2024)

# Get drought info for both counties in MORR in 2023
morr <- getClimDrought(park = "MORR", years = 2023, dom_county = TRUE)

}

}
