% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClimWStat.R
\name{getClimWStat}
\alias{getClimWStat}
\title{getClimWStat: Download daily data from nearest weather station}
\usage{
getClimWStat(park = "all", years = c(2006:2024), filepath = NA, export = FALSE)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD monitored for water quality.}
\item{"ACAD"}{Acadia NP only}
\item{"BOHA"}{Boston Harbor only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{years}{Vector of years to download weather station data for, will start with 01/01/year and end with 12/31/year.
Note that not all weather stations have complete a complete period of record from 2006 to current.}

\item{filepath}{Quoted path to save files to. If not specified, will save to working directory.}

\item{export}{Logical. If TRUE, will export a CSV of the compiled weather data with a date stamp. Must supply
a filepath to write output to. If FALSE (Default), will only return results to R environment.}
}
\value{
Data frame of weather station daily climate data for each specified park.
}
\description{
This function downloads daily weather data from data.rcc-acis.org from the nearest
weather station with the longest period of record dating through 2024 for each selected NETN park
based on its lat/long coordinates. Data include daily total precipitation (mm),max. temperature (C),
min temperature (C) and binds each site's data into a single data frame. Final data frame
can also be written to disk (export = T). If downloading for multiple parks and multiple years, function
may be slow, particularly ACAD. Note that the McFarland Hill weather station precip. data is not available
through typical weather station sources. Instead, hourly precip. data from NADP are downloaded and summarized
to daily value. ACAD NADP Precip data only go back to 2008.
}
\examples{
\dontrun{
# RUN FIRST
library(climateNETN)
importData()

# get weather data for ROVA from 2020-2023
rova <- getClimWStat(park = "ROVA", years = 2020:2023)

# get weather data for 2023 LNETN parks
lnetn <- getClimWStat(park = "LNETN", years = 2023)

}

}
