% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClimComps.R
\name{plotClimComps}
\alias{plotClimComps}
\title{plotClimComps: Plot climate comparisons}
\usage{
plotClimComps(
  park = "ACAD",
  years = 2006:format(Sys.Date(), "\%Y"),
  months = 1:12,
  layers = "lines",
  normal = "norm20cent",
  parameter = "tmean",
  units = "sci",
  plot_title = TRUE,
  palette = "viridis",
  color_rev = FALSE,
  legend_position = "right",
  gridlines = "none",
  ...
)
}
\arguments{
\item{park}{Specify park to plot. Currently can only plot 1 park at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"LNETN"}{Includes all parks but ACAD monitored for water quality.}
\item{"ACAD"}{Acadia NP only}
\item{"BOHA"}{Boston Harbor only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{years}{Numeric. Years to plot separately. Accepted values start at 2006.}

\item{months}{Vector of numeric months to query. If specifying new months not yet included in
NETN_clim_annual dataset, will download months that are available from NOAA.}

\item{layers}{Options are "points", "lines", or both for Annual Values. By default, only lines will plot.}

\item{normal}{Specify normal to plot. By default, the 20th century normal (1901-2000) plots.
Other options include:
\describe{
\item{"norm20cent"}{Plots the 20th century normal (1901 - 2000)}
\item{"norm1990"}{Plots the 30-year norm from 1991 - 2020}
}}

\item{parameter}{Specify the monthly averaged parameter to plot. Acceptable values are
\describe{
\item{"tmean"}{Plot mean temperature comparisons.}
\item{"tmax"}{Plot max temperature comparisons.}
\item{"tmin"}{Plot min temperature comparisons.}
\item{"ppt"}{Plot precipitation comparisons.}
}}

\item{units}{Specify if you want Scientific or English units. Acceptable values are "sci" (default) and "eng".
If "sci", temperature units are in C and precipitation units are in mm. If "eng", temperature units are in F,
and precipitation units are in inches.}

\item{plot_title}{Logical. If TRUE (default) prints site name at top of figure. If FALSE,
does not print site name. Only enabled when one site is selected.}

\item{palette}{Color palette for plots. Options currently are 'viridis' (yellow - green - blue),
or create your own by specifying 2 or more colors in quotes, like palette = c("red", "yellow", "blue"),
or palette = c("yellow", "blue"). Hexcodes work too. If only 1 year is specified, only need to
specify 1 color}

\item{color_rev}{Reverse the order of the color pallet. Only enabled for palette = 'viridis'.}

\item{legend_position}{Specify location of legend. To turn legend off, use legend_position =
"none" (Default). Other options are "top", "bottom", "left", "right".}

\item{gridlines}{Specify whether to add gridlines or not. Options are c("none" (Default), "grid_y", "grid_x", "both")}

\item{...}{Additional arguments relevant to \code{sumClimAvgs()} or \code{sumClimMonthly()}}
}
\value{
Returns a ggplot object of specified climate trends
}
\description{
This function compares historic average monthly climate variables with
user-specified years and months to provide an idea of how extreme or normal a given month in a
year is. This function only works with NOAA gridded climate data, and plots park
centroids, not site-level values. Each month's data are typically available about
two weeks after the month ends. To save download time, the NETN_clim_annual dataset contains
monthly climate data from January 2006 through May 2024.
}
\examples{
\dontrun{

# Plot mean monthly temp for MABI for 2019:2023 and all months with red-blue color palette
plotClimComps(park = "MABI", years = 2019:2023, parameter = "tmean", palette = c('red', 'blue'))

# Same as above, but with points too
plotClimComps(park = "MABI", years = 2019:2023, parameter = "tmean", palette = c('red', 'blue'),
  layers = c('points', 'lines'))

# Plot max monthly temp ACAD for 2019:2023 and all months with 1991- 2020 normals with english units
plotClimComps(park = "ACAD", years = 2019:2023, parameter = "tmax", normal = "norm1990",
  palette = c('red', 'blue'), layers = 'lines', units= 'eng')

# Plot total monthly precip for ACAD for past 10 years using a blue color scheme
plotClimComps(park = "ACAD", years = 2013:2023, parameter = 'ppt',
  palette = c("#75C5FF", "#3563DD", "#323969"))

# Plot total monthly precip for latest year in ACAD
plotClimComps(park = "ACAD", years = 2024, parameter = "ppt", months = 1:5)

}

}
