% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumStatsTable.R
\name{sumStatsTable}
\alias{sumStatsTable}
\title{sumStatsTable: Summarize climate records}
\usage{
sumStatsTable(
  park = "ACAD",
  years = 1895:format(Sys.Date(), "\%Y"),
  normal = "norm20cent",
  units = "sci",
  top_n = 5,
  kable = T
)
}
\arguments{
\item{park}{Specify park to plot. Currently can only plot 1 park at a time. Valid inputs:
\describe{
\item{"ACAD"}{Acadia NP only}
\item{"BOHA"}{Boston Harbor only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHP only}
\item{"SAIR"}{Saugus Iron Works NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHP only}}}

\item{years}{Numeric. Years to consider for summary table. Accepted values start at 1895.}

\item{normal}{Specify normal to plot. By default, the 20th century normal (1901-2000) plots.
Other options include:
\describe{
\item{"norm20cent"}{Plots the 20th century normal (1901 - 2000)}
\item{"norm1990"}{Plots the 30-year norm from 1991 - 2020}
}}

\item{units}{Specify if you want Scientific or English units. Acceptable values are "sci" (default) and "eng".
If "sci", temperature units are in C and precipitation units are in mm. If "eng", temperature units are in F,
and precipitation units are in inches.}

\item{top_n}{Number of rows to return per statistic. Default is 5.}

\item{kable}{Logical. If TRUE (default) will return a kable with formatting. If FALSE, will return a data.frame.}
}
\value{
Returns a dataframe or kable of climate statistics
}
\description{
This function summarizes climate records by month, including wettest, dryest,
warmest and coolest based on monthly values dating back to 1895 for a given park.
}
\examples{
\dontrun{

# Generate html table of top 5 records in ACAD for all years in English
sumStatsTable(park = "ACAD", units = "eng", top_n = 5, kable = T)

# Return dataframe of top 3 records in WEFA for all years in scientific units
sumStatsTable(park = "WEFA", top_n = 3, kable = F)

# Generate
}

}
